+++ src/Event/EventList.php
@@ -16,6 +16,7 @@ declare(strict_types=1);
 namespace Tools\Event;
 
 use ArrayAccess;
+use Tools\Event\Event;
 
 /**
  * EventList.
@@ -26,7 +27,7 @@ class EventList implements ArrayAccess
 {
     /**
      * Events list
-     * @var array<\Tools\Event\Event>
+     * @var \Tools\Event\Event[]
      */
     protected array $_events = [];
 
@@ -55,7 +56,7 @@ class EventList implements ArrayAccess
      * @param mixed $offset An offset to check for
      * @return bool True on success or false on failure
      */
-    public function offsetExists($offset): bool
+    public function offsetExists(mixed $offset): bool
     {
         return isset($this->_events[$offset]);
     }
@@ -64,9 +65,9 @@ class EventList implements ArrayAccess
      * Offset to retrieve
      * @link https://secure.php.net/manual/en/arrayaccess.offsetget.php
      * @param mixed $offset The offset to retrieve
-     * @return mixed
+     * @return \Tools\Event\Event|null
      */
-    public function offsetGet($offset)
+    public function offsetGet(mixed $offset): ?Event
     {
         return $this->offsetExists($offset) ? $this->_events[$offset] : null;
     }
@@ -78,7 +79,7 @@ class EventList implements ArrayAccess
      * @param mixed $value The value to set
      * @return void
      */
-    public function offsetSet($offset, $value): void
+    public function offsetSet(mixed $offset, mixed $value): void
     {
         $this->_events[$offset] = $value;
     }
@@ -89,7 +90,7 @@ class EventList implements ArrayAccess
      * @param mixed $offset The offset to unset
      * @return void
      */
-    public function offsetUnset($offset): void
+    public function offsetUnset(mixed $offset): void
     {
         unset($this->_events[$offset]);
     }
@@ -106,7 +107,7 @@ class EventList implements ArrayAccess
     /**
      * Extracts events by name
      * @param string $name Event name
-     * @return array<\Tools\Event\Event>
+     * @return \Tools\Event\Event[]
      * @since 1.5.12
      */
     public function extract(string $name): array
@@ -126,7 +127,7 @@ class EventList implements ArrayAccess
 
     /**
      * Returns the `EventList` as array
-     * @return array<\Tools\Event\Event>
+     * @return \Tools\Event\Event[]
      * @since 1.4.1
      */
     public function toArray(): array