+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 1.9 branch
 ### 1.9.0
 * requires at least PHP 8.1. Updated `cakephp-codesniffer`, `phpstan` and `psalm` packages;
+* updated `Event` and `EventList` classes;
 * added tests for PHP 8.2 and 8.3.
 
 ## 1.8 branch